Added SMS notification method. Added stress test program. Refactroring
